Move check to Config
